requirements.txt
google-generativeai
gradio            
Pillow             
requests          
python-dotenv      